pyproject.toml
[tool.poetry]
name = "mindexpo"
version = "0.1.0"
description = ""
authors = ["Bikash Jaiswal <bikash.jaiswal@cerence.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
ruff = "^0.0.278"
markdown = "^3.4.3"
jinja2 = "^3.1.2"
python-frontmatter = "^1.0.0"
email-validator = "^2.0.0.post2"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "mindexpo.main:start" 